Made changes to global_logo.pt.  Fixed the problem with the college logo in the undergraduate site point.  Eric Steele suggested using portal_url.  This takes you to the root of the current Plone site on both the test box and the live site without changing paths.
Made changes to background color for portlet_navigation in portlets.css.  Color is now EEEECC so it matches the surrounding background.  

.navTreeLevel2 { 
padding-left:1em;
margin-left: 0; 
background-color: #eeeecc;
}

It was #ddb

Also changed the color of the fonts 
color for the un-highted text is #808183  it was 
.navTreeLevel2 a {
/*color:#006;*/
color: #808183;
}

color for the highlighted text is #022b54  This is globalFontColor:string=#022b54 in base_properties.props    It was #436976